General structure and basic parser
This lays the groundwork for a basic parser for SGQL.  Right now, there
is the ability to run a simple SELECT query with the location graph.
The parser will catch errors and throw an exception with details of what
it was expecting / missing, and at what location.  In addition, all of
the necessary parsing rules have unit tests.