Change behavior of HAS token
HAS is now a function, and behaves much like the aggregation and count
functions.  Since HAS is a check on a set of records, it makes more
sense to use it as a function in the parent record of that set, rather
than on the set itself.

In a simple example, a customer can have many orders associated with
them.  If writing a query to return a customer and their orders only if
that customer has order ID 5, the HAS function can be used in the
customers namespace as such:

`customers:(HAS(orders:id == 5) == true)`

This doesn’t filter any sets of orders, it filters customers based on
their orders.  This aligns with the purpose of a function such as COUNT,
which counts child records in the parent.  The new HAS function checks
if a child record exists in a given namespace.