Add mechanism for database connections
Database holds the connection info for all hosts, which are each an
individual user login to a database.  This allows for switching between
databases and handles all of the details.  The config array passed in
uses regex to determine which users have access to each database, and
how to connect to them.