Abstract database driver class
This class is the abstract definition of a database driver, which is
responsible for making queries against the database.  Each database will
have its own class that extends this, which will facilitate differences
in the APIs used to connect to each database, but provide only one API
to the code in this project.