Major change to graph storage
This transitions from a program-supplied graph config to automated graph
storage and retrieval.  This stores the graph configuration in 3 tables
in each SGQL enabled database: sql_info, sql_tables,
and sgql_associations.  This touches many parts of code due to the fact
that the way the graph is constructed relies solely on the database
driver, and does not require any programmer input to build a graph or
store its configuration.