Don’t allow primary columns to be aliased
If this happens, we have no way of knowing how to associate child
records, because we might not know what the alias is for the primary
column used to associate them.  We could probably figure this out, but
it isn’t good practice to alias the primary column, and it’s easier
this way.