Create association table for associations
This creates the actual mapping table for individual associations.  This
will map a parent record to a child record, and store the primary column
values for both.  In the Association class, there are methods to
determine which column is for which schema, as well as find out what
the table name is for the association.