Allow construction of insert queries
While execution is still not available, this allows the user to build
insert queries.  `insert` takes 2 parameters - `$values` and
`$associations`.  `$values` is the only parameter used at the moment,
and allows the user to pass in namespaced values to insert, and
automatically validates and transforms them into a format that the
executor will be able to use.

If a string query is used to insert, there are transformer methods that
handle the transformation to the structure needed for `insert`, and then
`insert` will validate and transform the values again into a format
understood by the executor.