Add function to check for consecutive IDs
When bulk inserting, it is important that we are guaranteed that the IDs
will be consecutive.  Otherwise, there might be holes, which would break
future insert code that relies on consecutive IDs with no holes (in a
single statement).  So, before allowing the insert function to finish,
we check that the database is configured to set consecutive IDs.