Add insert queries
This adds the remaining code necessary for basic insert queries.  This
allows the user to have insert nested values and automatically takes
care of the association, and returns an array that has the updated
primary column IDs for each value.

One outstanding TODO: When creating associations, we need to check that
the values being associated do not violate the rules of the specific
association.

This does not add any code for associating records in the top level
schema to other records.